+++ lib/shared/common/page_mode_enum.dart
@@ -1,0 +1,12 @@
+/// lib/shared/common/page_mode_enum.dart
+///
+/// a enum to indicate the page mode
+///
+/// - [view]: the page is in view mode
+/// - [edit]: the page is in edit mode
+///
+/// useful to limited user's action on the page
+enum PageMode {
+  view,
+  edit,
+}